+++ src/components/admin/sidebar/sidebar.css
@@ -4,6 +4,7 @@
   background-color: rgb(251, 251, 255);
   position: sticky;
   top: 50px;
+  height:100%,
 }
 
 .sidebarWrapper {
