+++ src/pages/admin/productList/ProductList.jsx
@@ -8,12 +8,18 @@ import { Box, Button, FormControl, InputLabel, MenuItem, Modal, Select, Stack, T
 import Confirmation from "../../../components/admin/alert/Confirmation";
 import AddIcon from '@mui/icons-material/Add';
 import CloseIcon from '@mui/icons-material/Close';
+import CategoryList from "../categoryList/categoryList";
 
 const Container = styled.div`
-    padding: 20px;
+    padding: 5px;
     display: flex;
     flex-wrap: wrap;
     justify-content: space-between;
+`;
+
+const Title = styled.h2`
+  margin: 10px;
+  font-weight: normal;
 `;
 
 
@@ -55,7 +61,8 @@ class ProductList extends React.Component{
         img_url: null,
         categoryId: '',
         editOpen: false,
-        openImage: false
+        openImage: false,
+        pagination:5
     }
   }
 
@@ -65,7 +72,7 @@ class ProductList extends React.Component{
   }
 
   getProductList = () => {
-    Axios.get('http://localhost:5000/api/products')
+    Axios.get(process.env.REACT_APP_API+'/admin/products')
     .then((response) => {
         this.setState({dbProduct: response.data})
     })
@@ -87,7 +94,7 @@ class ProductList extends React.Component{
   onButtonConfirmDelete = () => {
     this.setState({alertDelete: false});
 
-    Axios.delete(`http://localhost:5000/api/products/${this.state.deleteId}`)
+    Axios.delete(process.env.REACT_APP_API + `/admin/products/${this.state.deleteId}`)
     .then((response) => {
       alert('Successfully Deleted!');
       this.getProductList();
@@ -122,11 +129,11 @@ class ProductList extends React.Component{
    
     console.log(formData);
 
-    Axios.post('http://localhost:5000/api/products', formData)
+    Axios.post(process.env.REACT_APP_API+`/admin/products/${this.state.editId}`, formData)
     .then((response) => {
-        console.log("Successfully Added New Category!");
+        console.log("Successfully Added New Product!");
         this.setState({addOpen: false});
-        alert('Successfully Added New Category!');
+        alert('Successfully Added New Product!');
         this.setState({name: null})
         this.setState({price: null})
         this.setState({weight_gram: null})
@@ -161,7 +168,7 @@ class ProductList extends React.Component{
       formData.append("file", this.state.img_url);
       formData.append("categoryId", this.state.categoryId);
       // {name: this.state.name, price: this.state.price, weight_gram:this.state.weight_gram, description: this.state.description, file: this.state.img_url, categoryId: this.state.categoryId}
-      Axios.patch(`http://localhost:5000/api/products/${this.state.editId}`, formData)
+      Axios.patch(process.env.REACT_APP_API+`/admin/products/${this.state.editId}`, formData)
       .then((response) => {
           alert('Successfully Updated Product');
           this.setState({editOpen: false});
@@ -201,13 +208,13 @@ class ProductList extends React.Component{
   }
 
   getCategories = () => {
-    Axios.get('http://localhost:5000/api/categories')
+    Axios.get(process.env.REACT_APP_API+'/admin/categories')
     .then((response) => {
         this.setState({dbCategory: response.data})
     })
     .catch(err => {
       console.log(err);
-  })
+    })
   }
   
   printCategory = () => {
@@ -274,13 +281,21 @@ class ProductList extends React.Component{
     
     return (
       <Container>
+        <Box display="flex" flexDirection="column">
+          <Box sx={{marginBottom: '20px'}}>
+            <Title>Catogory Product</Title>
+            <CategoryList/>
+          </Box>
+          <Box>
+              <Button size="small" onClick={() => this.btnCancel()}><AddIcon/>Add Products</Button>
+          </Box>
+        </Box>
         <Stack direction="row" spacing={1} sx={{ mb: 1 }}>
-          <Button size="small" onClick={() => this.btnCancel()}><AddIcon/>Add Products</Button>
           <Modal
             open={this.state.addOpen}
           >
             <Box sx={style}>
-                <Box position='absolute' right='5%'>
+            <Box position='absolute' right='5%'>
                   <CloseBtn onClick={()=> this.setState({addOpen: false})}><CloseIcon sx={{color: '#008080'}}/></CloseBtn>
                 </Box>
                 <Box marginTop='30px'>
@@ -290,51 +305,63 @@ class ProductList extends React.Component{
                   <br/>
                   <Stack spacing='20px'>
                     
-                    <TextField fullWidth label="Product Name" id="name"
-                      onChange ={(event) => this.handleUserInput(event)}
-                      value = {this.state.name}
-                      size='small'
-                      name="name"
-                    />
+                    <FormControl>
+                      <TextField fullWidth label="Product Name" id="name"
+                        onChange ={(event) => this.handleUserInput(event)}
+                        value = {this.state.name}
+                        size='small'
+                        name="name"
+                        required
+                      />
+                    </FormControl>
 
-                   
-                    <TextField fullWidth label="Price" id="price"
-                      onChange ={(event) => this.handleUserInput(event)}
-                      value = {this.state.price}
-                      size='small'
-                      name="price"
-                    />
+                    <FormControl>
+                        <TextField fullWidth label="Price" id="price"
+                          onChange ={(event) => this.handleUserInput(event)}
+                          value = {this.state.price}
+                          size='small'
+                          name="price"
+                          required
+                        />
+                    </FormControl>
                     
-                    
-                    <TextField fullWidth label="Weight(Gram)" id="weight_gram"
-                      onChange ={(event) => this.handleUserInput(event)}
-                      value = {this.state.weight_gram}
-                      size='small'
-                      name="weight_gram"
-                    />
+                    <FormControl>
+                      <TextField fullWidth label="Weight(Gram)" id="weight_gram"
+                        onChange ={(event) => this.handleUserInput(event)}
+                        value = {this.state.weight_gram}
+                        size='small'
+                        name="weight_gram"
+                        required
+                      />
+                    </FormControl>
 
-                   
-                    <TextField fullWidth label="Description" id="description"
-                      onChange ={(event) => this.handleUserInput(event)}
-                      value = {this.state.description}
-                      size='small'
-                      name="description"
-                    />
-                    
-                    <InputLabel id="category-select">Category</InputLabel>
-                    <Select
-                      labelId="category-select"
-                      id="categoryId"
-                      value={this.state.categoryId}
-                      label="Category"
-                      onChange={(event) => this.handleUserInput(event)}
-                      name="categoryId"
-                      size='small'
-                    >
-                       {this.printCategory()}
-                    </Select>
+                   <FormControl>
+                      <TextField fullWidth label="Description" id="description"
+                        onChange ={(event) => this.handleUserInput(event)}
+                        value = {this.state.description}
+                        size='small'
+                        name="description"
+                        required
+                      />
+                   </FormControl>
 
-                    <InputLabel id="category-select">Upload Image</InputLabel>
+                    <FormControl>
+                      <InputLabel id="category-select">Category</InputLabel>
+                      <Select
+                        labelId="category-select"
+                        id="categoryId"
+                        value={this.state.categoryId}
+                        label="Category"
+                        onChange={(event) => this.handleUserInput(event)}
+                        name="categoryId"
+                        size= 'medium'
+                        required
+                      >
+                        {this.printCategory()}
+                      </Select>
+                    </FormControl>
+
+                    <InputLabel id="category-select" required>Upload Image</InputLabel>
                     <Box color='neutral' display='flex' flex-direction='row' justifyContent='center' alignItems='center' width='100%' height='100px' border='1px dotted #508dcd'>
                       <label htmlFor="productMenu">
                         <Input accept="image/*" id="productMenu" multiple type="file" onChange={(e) => this.setState({img_url: e.target.files[0]})}/>
@@ -427,9 +454,15 @@ class ProductList extends React.Component{
             </Box>
           </Modal>
         </Stack>
-        <div style={{ height: 400, width: '100%' }}>
+        <div style={{ height: 350, width: '100%' }}>
           <Confirmation isOpen={this.state.alertDelete} title="Confirmation Delete" onCancel={this.onButtonCancelDelete} onConfirm={this.onButtonConfirmDelete}/>
-          <DataGrid rows={this.state.dbProduct} columns={columns} components={{ Toolbar: GridToolbar }} />
+          <DataGrid 
+            rows={this.state.dbProduct} 
+            columns={columns} 
+            components={{ Toolbar: GridToolbar }} 
+            pageSize={3}
+            rowsPerPageOptions={[3]} 
+          />
         </div>
       </Container>
     );
